added models to seperate file
